SunOS: silence some warnings
Give `-Wformat=0` on illumos/Solaris so that the build does not
complain about `execlp` not having a sentinel (it's nil).  Also
fix an integer size issue when casting to a void pointer.

+++ src/libthread/thread.c
@@ -908,7 +908,7 @@ threadinfo(void *v, char *s)
 		return 0;
 
 	for(p=_threadprocs; p; p=p->next){
-		fprint(2, "proc %p %s%s\n", (void*)p->osprocid, p->msg,
+		fprint(2, "proc %p %s%s\n", (void*)(uintptr)p->osprocid, p->msg,
 			p->sysproc ? " (sysproc)": "");
 		for(t=p->allthreads.head; t; t=t->allnext){
 			fprint(2, "\tthread %d %s: %s %s\n",
